Add arc plotting function, add 3D spline plotting
Added vcg_analysis.plot_arc3d - allows to plot a filled arc between
two vectors, representing the angular difference between the two of
them.

Added vcg_analysis.add_xyz_axes to replace default axes in 3D plot
with self-constructed axes to make it more visually appealing. Note
that this function must be called explicitly, and it is recommended
that it is called at the end to ensure that over-lapping dummy axes
are not created.

Added vcg_analysis.add_unit_sphere, moving the plotting of a unit
sphere in 3D space to that function. Note that adding a unit sphere
is no longer done by default, and an explicit call must be made.