Add function to plot all metric data
Added function to plot all metric analysis in one figure, rather
than requiring it to be combined in Inkscape or other later. Will
plot scatter plots of a metric change compared to volume and
surface area of scar, and line plots for how it changes in LV and
septum, depending on changes in phi, rho and z.

Also some tidying up. add_unit_sphere is now designed to be called
separately.