Allow direct passing of time variable instead of dt/t_end
Adjusted vcg_analysis and vcg_plot to allow time variable to be
passed directly to the analysis/plotting functions, rather than
having the time variable be calculated based on assumed dt/t_end
variables. As part of this, incorporated function in
common_analysis to parse the various data functions and give
unified outputs.