Add get_i_colour, enhance input_to_list
Added the get_i_colour function to tools_python, which gets the number
of lines already extant on a plot (and thus permits relatively easy
colouring of subsequent lines subject to any required colour scheme).

Added option to not perform length checks on input arguments when
converting to list, i.e. just ensure that the lists are of the required
depth, without making any other requirements.