Update set_axis_limits,add_colourbar,add_xyz_axes
Update tools_plotting.set_axis_limits to allow data to be an optional
argument, with the update to reflect that it should be pd.DataFrame
rather than np.ndarray.

Update tools_plotting.add_colourbar with reference to figure included in
possible arguments, and added default arguments.

Import add_xyz_axes to tools_plotting, and improved to allow colourbar
to be retained as axis is shifted to magnify image better.